% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upipe.R
\name{upipe}
\alias{\%>\%}
\alias{upipe}
\title{Pipe through system shell}
\usage{
upipe(input, command, show_stderr = interactive())

input \%>\% upipe(command)
}
\arguments{
\item{input}{a character string to send to the shell. Elements will be separated
by newlines.}

\item{command}{a system command to take \code{input} as stdin}

\item{show_stderr}{print stderr of the command to the console as a message?}
}
\description{
\code{upipe} is a modification to \code{\link[base]{system}} designed for
piping data out of R, into the shell, and returning back to an R analysis pipeline
Take in a text object, run it through a unix command, return it
}

